Add update functionality for WASM code id
The commit introduces the new rpc handler method, `UpdateWasmCodeId`, for updating the WASM code id within IBC light clients. This allows clients to change the existing code id for WASM contracts. The new function takes in three parameters, signer, client id, and code id and updates the existing code id for the given client. This functionality was necessary for providing flexibility in managing the contracts.

The update can only be triggered by the authority.